Remaining issue to solve: With no treatment the C++ ratios between the Infant and Pre-SAC, Pre-SAC and SAC, SAC and adult mean female worm numbers are similar to those of R. However, in C++, the female worm numbers keep decreasing over time instead of stabilising. Find out why.

+++ CHost.h
@@ -18,7 +18,7 @@ public:
 	// Members.
 	double birthDate, deathDate;
 	double femaleWorms, totalWorms;
-	int hostAge;
+	double hostAge;
 };
 
 #endif /* CHOST_H_ */
